requirements.txt
future==0.16.0
requests==2.18.4
PyExecJS==1.5.0
